+++ agentrule-maven-plugin/src/main/java/io/opentracing/contrib/specialagent/FingerprintMojo.java
@@ -17,6 +17,7 @@ package io.opentracing.contrib.specialagent;
 
 import java.io.File;
 import java.io.IOException;
+import java.lang.reflect.Field;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLClassLoader;
@@ -26,7 +27,6 @@ import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.DefaultArtifact;
 import org.apache.maven.artifact.handler.ArtifactHandler;
 import org.apache.maven.artifact.repository.ArtifactRepository;
-import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.Execute;
@@ -34,7 +34,7 @@ import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
-import org.apache.maven.project.MavenProject;
+import org.apache.maven.plugins.dependency.tree.TreeMojo;
 import org.codehaus.plexus.component.repository.ComponentDependency;
 
 import io.opentracing.contrib.specialagent.Link.Manifest;
@@ -55,9 +55,9 @@ import io.opentracing.contrib.specialagent.Link.Manifest;
  * and class types).</li>
  * </ol>
  */
-@Mojo(name="fingerprint", defaultPhase=LifecyclePhase.GENERATE_SOURCES, requiresDependencyResolution=ResolutionScope.TEST)
+@Mojo(name="fingerprint", defaultPhase=LifecyclePhase.PROCESS_CLASSES, requiresDependencyResolution=ResolutionScope.TEST)
 @Execute(goal="fingerprint")
-public final class FingerprintMojo extends AbstractMojo {
+public final class FingerprintMojo extends TreeMojo {
   /**
    * Returns an {@code Artifact} representation of {@code dependency}, qualified
    * by {@code artifactHandler}.
@@ -136,32 +136,39 @@ public final class FingerprintMojo extends AbstractMojo {
     return depth == 0 ? null : new URL[depth];
   }
 
-  @Parameter(defaultValue="${project}", required=true, readonly=true)
-  private MavenProject project;
-
-  @Parameter(defaultValue="${localRepository}")
+  @Parameter(defaultValue="${localRepository}", required=true, readonly=true)
   private ArtifactRepository localRepository;
 
-  @Parameter(defaultValue="${sa.plugin.name}")
+  @Parameter(defaultValue="${sa.plugin.name}", required=true, readonly=true)
   private String name;
 
-  @Override
-  public void execute() throws MojoExecutionException, MojoFailureException {
-    if ("pom".equalsIgnoreCase(project.getPackaging()))
-      return;
-
-    if (name == null || name.length() == 0)
-      throw new MojoExecutionException("The parameter 'name' is missing or invalid");
-
+  private void setField(final Class<? super FingerprintMojo> cls, final String fieldName, final Object value) {
     try {
-      final File destFile = new File(project.getBuild().getOutputDirectory(), "fingerprint.bin");
+      final Field field = cls.getDeclaredField(fieldName);
+      field.setAccessible(true);
+      field.set(this, value);
+    }
+    catch (final IllegalAccessException | NoSuchFieldException e) {
+      throw new IllegalStateException(e);
+    }
+  }
+
+  private void createDependenciesTgf() throws MojoExecutionException, MojoFailureException {
+    setField(TreeMojo.class, "outputType", "tgf");
+    setField(TreeMojo.class, "outputFile", new File(getProject().getBuild().getOutputDirectory(), "dependencies.tgf"));
+    super.execute();
+  }
+
+  private void createFingerprintBin() throws MojoExecutionException, MojoFailureException {
+    try {
+      final File destFile = new File(getProject().getBuild().getOutputDirectory(), RuleClassLoader.FINGERPRINT_FILE);
       destFile.getParentFile().mkdirs();
       final File nameFile = new File(destFile.getParentFile(), "sa.plugin.name." + name);
-      if (!nameFile.createNewFile())
+      if (!nameFile.exists() && !nameFile.createNewFile())
         throw new MojoExecutionException("Unable to create file: " + nameFile.getAbsolutePath());
 
       // The `optionalDeps` represent the 3rd-Party Library that is being instrumented
-      final URL[] optionalDeps = getDependencyPaths(localRepository, null, true, project.getArtifacts().iterator(), 0);
+      final URL[] optionalDeps = getDependencyPaths(localRepository, null, true, getProject().getArtifacts().iterator(), 0);
       if (optionalDeps == null) {
         getLog().warn("No dependencies were found with (scope=*, optional=true), " + RuleClassLoader.FINGERPRINT_FILE + " will be empty");
         new LibraryFingerprint().toFile(destFile);
@@ -170,20 +177,32 @@ public final class FingerprintMojo extends AbstractMojo {
 
       // The `compileDeps` represent the Instrumentation Plugin (this is the dependency(ies)
       // that bridges/links between the 3rd-Party Library to the Instrumentation Rule).
-      final URL[] compileDeps = getDependencyPaths(localRepository, "compile", false, project.getArtifacts().iterator(), 1);
+      final URL[] compileDeps = getDependencyPaths(localRepository, "compile", false, getProject().getArtifacts().iterator(), 1);
       // Include the compile path of the Instrumentation Rule itself, which solves the use-
       // case where there is no Instrumentation Plugin (i.e. the Instrumentation Rule directly
       // bridges/links between the 3rd-Party Library to itself).
-      compileDeps[0] = new File(project.getBuild().getOutputDirectory()).toURI().toURL();
+      compileDeps[0] = new File(getProject().getBuild().getOutputDirectory()).toURI().toURL();
 
       final Manifest manifest = Link.createManifest(compileDeps);
 
-      final URL[] nonOptionalDeps = getDependencyPaths(localRepository, null, false, project.getArtifacts().iterator(), 0);
+      final URL[] nonOptionalDeps = getDependencyPaths(localRepository, null, false, getProject().getArtifacts().iterator(), 0);
       final LibraryFingerprint fingerprint = new LibraryFingerprint(new URLClassLoader(nonOptionalDeps), manifest, optionalDeps);
       fingerprint.toFile(destFile);
     }
     catch (final IOException e) {
       throw new MojoFailureException(null, e);
     }
+  }
+
+  @Override
+  public void execute() throws MojoExecutionException, MojoFailureException {
+    if ("pom".equalsIgnoreCase(getProject().getPackaging()))
+      return;
+
+    if (name == null || name.length() == 0)
+      throw new MojoExecutionException("The parameter 'name' is missing or invalid");
+
+    createDependenciesTgf();
+    createFingerprintBin();
   }
 }